+++ src/main/java/subway/line/dao/LineDao.java
@@ -1,10 +1,11 @@
-package subway.line;
+package subway.line.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import subway.line.domain.Line;
 
 import java.sql.PreparedStatement;
 import java.util.List;
