+++ src/main/java/subway/station/service/StationService.java
@@ -1,6 +1,8 @@
-package subway.station;
+package subway.station.service;
 
 import org.springframework.stereotype.Service;
+import subway.station.dao.StationDao;
+import subway.station.domain.Station;
 
 import java.util.List;
 import java.util.stream.Collectors;
